Added an install script to the MagellanTools folder. This will: -check for all modules -create directories -copy over files -set permissions -create database.dat File -create login table
I put together my proposed packaging method in the rc-example folder. The zip is what we would put online, the folder is just an example of what is in the zip. This needs to be updated to work with Caleb's latest commit which adds a new file, that will happen tomorrow. Also There may be an uninstall script in the future.